% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_bz_data.r
\name{create_bz_data}
\alias{create_bz_data}
\title{Run the Shiny Application}
\usage{
create_bz_data(conn, project = 1916, pers_set = 79431,
  nieuws_set = 79457, ...)
}
\arguments{
\item{conn}{An amcatr connection}

\item{project}{AmCAT project}

\item{pers_set}{AmCAT articleset with BZ press articles}

\item{nieuws_set}{AmCAT articleset with BZ news articles}

\item{...}{}
}
\description{
Run the Shiny Application
}
