% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_viewer.R
\name{mod_viewer_ui}
\alias{mod_viewer_ui}
\alias{mod_viewer_server}
\title{mod_viewer_ui and mod_viewer_server}
\usage{
mod_viewer_ui(id)

mod_viewer_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
