% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_bz_data.r
\name{prepare_data}
\alias{prepare_data}
\title{Prepare similarity data}
\usage{
prepare_data(tc, from_subset, to_subset, date_col, ...)
}
\arguments{
\item{tc}{a tCorpus}

\item{from_subset}{a subset expression for selecting 'from' which documents to compare}

\item{to_subset}{a subset expression for selecting 'to' which documents to compare}

\item{date_col}{The column in tc$meta that contains a date (has to be in POSIXct)}

\item{...}{Optionally, default values for min_similarity, hour_window, ngrams, measure and weight}
}
\description{
Prepare similarity data
}
